style: fix additional mypy type issues
- Fix dict type annotation from int to float in ib_config.py chunk_days
- Add Optional type hints for function parameters to fix PEP 484 compliance:
  - model_metadata.py: asset_class parameters in compatibility methods
  - config_builder.py: panels parameter in create_multi_panel_config
- Fix Optional type annotations for dataclass fields in model_metadata.py:
  - training_data, deployment_capabilities, feature_architecture
  - performance_metrics, training_config fields

These fixes address common type annotation patterns and resolve explicit
Optional type requirements under strict mypy configuration.

+++ ktrdr/config/ib_config.py
@@ -40,7 +40,7 @@ class IbConfig:
     )
 
     # Data fetching settings
-    chunk_days: dict[str, int] = field(
+    chunk_days: dict[str, float] = field(
         default_factory=lambda: {
             "1 secs": 0.02,  # 30 minutes
             "5 secs": 0.08,  # 2 hours
